# Conflicts:
#	.github/workflows/PullRequest-DailyJobs.yml
#	.github/workflows/PullRequest-HourlyJobs.yml
#	.github/workflows/PullRequest-WebApp.yml
#	.github/workflows/container_ca-tm-dev-westus2.yml
#	CLAUDE.md
#	Planning/Projects/Project_06_Backend_Standards.md
#	Planning/Projects/Project_27_OpenTelemetry_Migration.md
#	TrashMob/Common/TrashMobActivitySources.cs
#	TrashMob/Controllers/AdminController.cs
#	TrashMob/Controllers/ContactRequestController.cs
#	TrashMob/Controllers/EventAttendeeRoutesController.cs
#	TrashMob/Controllers/EventAttendeesController.cs
#	TrashMob/Controllers/EventLitterReportController.cs
#	TrashMob/Controllers/EventSummariesController.cs
#	TrashMob/Controllers/EventsController.cs
#	TrashMob/Controllers/JobOpportunityController.cs
#	TrashMob/Controllers/LitterReportController.cs
#	TrashMob/Controllers/PartnerAdminInvitationsController.cs
#	TrashMob/Controllers/PartnerAdminsController.cs
#	TrashMob/Controllers/PartnerContactsController.cs
#	TrashMob/Controllers/PartnerController.cs
#	TrashMob/Controllers/PartnerDocumentsController.cs
#	TrashMob/Controllers/PartnerLocationContactsController.cs
#	TrashMob/Controllers/PartnerLocationServicesController.cs
#	TrashMob/Controllers/PartnerLocationsController.cs
#	TrashMob/Controllers/PartnerRequestsController.cs
#	TrashMob/Controllers/PartnerSocialMediaAccountsController.cs
#	TrashMob/Controllers/PickupLocationsController.cs
#	TrashMob/Controllers/SecureController.cs
#	TrashMob/Controllers/UsersController.cs
#	TrashMob/Program.cs
#	TrashMob/TrashMob.csproj
#	TrashMob/client-app/package-lock.json
#	TrashMob/client-app/src/App.tsx
#	TrashMob/client-app/src/pages/faq/page.tsx
#	TrashMobMobile/Authentication/AuthConstants.cs
#	TrashMobMobile/TrashMobMobile.csproj

+++ TrashMob.Models/Poco/Stats.cs
@@ -23,12 +23,12 @@
         /// <summary>
         /// Gets or sets the total weight of litter collected in pounds.
         /// </summary>
-        public int TotalWeightInPounds { get; set; }
+        public decimal TotalWeightInPounds { get; set; }
 
         /// <summary>
         /// Gets or sets the total weight of litter collected in kilograms.
         /// </summary>
-        public int TotalWeightInKilograms { get; set; }
+        public decimal TotalWeightInKilograms { get; set; }
 
         /// <summary>
         /// Gets or sets the total number of participants across all events.
